Implement basic saving to file functionality.
Note that I circumvented RichTextFX's own codecs, specifically its listCodec() because it always writes the size of the list to the file in binary format, which will result in a mixed-binary-plain-text file.
Also, the codec API is pretty rigid as it always expects a DataOutputStream. DataOutputStreams write to a file in binary format, which is not what is wanted here, so I had to implement a static BufferedWriter to circumvent it.