Fix: closing comments does not end comment style (#62)
Several things were done in this commit:

- for added structure, all extension methods operating on strings or
char sequences have been extracted into a new file StringExtensions.kt
- the caret position listener now checks if the caret is at the position
right after a comment end token and, if so, applies an "empty" text
insertion style. It feels like there should be a better way to handle
token-delimited styles, but it does work for now.
- I completely overhauled the updateStyles() method and renamed it
updateComments(). It now requires a PlainTextChange to be passed in and
no longer updates the comment styles on the entire document but only on
the subsection that was actually edited (according to the change). This
required the addition of a new style span mutator method that works
opposite the old mergeStyles() (now unionStyles) to detect when a token
was removed. Previously all styles were removed and then re-added; now
styles are always maintained and only removed if a token was actually
removed.