Move unchecked casts to Controls.kt
Previously, there were unchecked casts from Property<Double> to
Property<Number> in the EditThemeView. The cast was necessary because
the property generic is invariant - and it has to be, as it serves as
both an argument (in) as well as return type (out).

However, since all number types inherit from Number and all Numbers have
extension methods to convert it to the desired concrete type (like Int),
there is no harm in pretending Property<T> is bivariant.

To reduce unnecessary boilerplate code, the casts from Property<Double>
to Property<Number> have been moved from EditThemeView to the pixelfield
declaration in ControlsKt.