Improve theme preview scaling
Previously the theme preview used values relative to the sizes of the screen and, in some cases, the main window of Skywriter. This makes little sense, however, and resulted in inconsistent scaling when the main window was small versus when it was large.

This has now been corrected. There's now a global SCALING_FACTOR that is applied to any element without a definitive relative parent element. Examples include the padding. Other elements, like the document width and height, are relative to the screen only, and everything is additionally relative to a preview's "original" size (i.e. the predefined size that the container starts at when the window is initially opened).