Update selectWords/selectLines logic
Previously selectWords and selectLines tried to derive their new start and end points based on the caret and anchor positions. However, the anchor position is unreliable -- it is always smaller than or equals to the caret position and thus provides inconsistent feedback.

My first thought was to use the selection range instead, but this also proved unreliable as a user would only be able to expand a selection, not shrink it. To shrink it, the information of what the original hit point is was necessary. The new implementation stores the index of the original user's mouse click in the text area and uses only the caret to determine the new selection range, ignoring the anchor. This is consistent and reduced code by more than 50%.

All this was necessary because I wanted the word selection logic to use the WordBreakIterator, as it does in other places. (Previously it just used an extension method to split the string and find its start and end points.) Apostrophes (including smart apostrophes) are now properly considered part of a word.

Technically, selectLines() does not use the WordBreakIterator and thus would have worked fine without a rework. However, for the sake of consistency, selectLines() now also uses an IntRange as its parameter which is constructed of the caret position and origin hit point.