Restructure WriterTextArea to reduce class size
To prevent the WriterTextArea from growing into a "god class", the following functionality has been split into separate files and added to the WriterTextArea using extension methods:

- actions pertaining to navigation/selection/editing -> TextAreaNavigator
- actions pertaining to styles (retrieving, setting) -> TextAreaStyler
- the context menu -> WriterViewContextMenu

Originally, I wanted to use something other than extension methods because I considered extension methods the "quick and dirty" way to accomplish what I wanted, but not exactly the ideal way. However, as all the moved methods require access to the WriterTextArea as their context, only two options remained:

- use composition in the same vein as the SmartReplacer or the TextAreaSearcher, but this would have required all calls to be prefixed by a navigation.* or styling.*, which would have been very unclean or
- use extension methods.

Kotlin supports a feature called class delegation, but class delegation does not allow you to pass `this` as the context to the delegation class since the context class has not been constructed yet at the time of the delegation class construction. If not for that, it would have been the ideal solution.